+++ src/paas_charm/charm_state.py
@@ -10,7 +10,11 @@ from dataclasses import dataclass, field
 
 from pydantic import BaseModel, Field, ValidationError, create_model
 
-from paas_charm.exceptions import CharmConfigInvalidError, InvalidRelationDataError
+from paas_charm.exceptions import (
+    CharmConfigInvalidError,
+    InvalidRelationDataError,
+    RelationDataError,
+)
 from paas_charm.secret_storage import KeySecretStorage
 from paas_charm.utils import build_validation_error_message, config_metadata
 
@@ -18,8 +22,10 @@ from paas_charm.utils import build_validation_error_message, config_metadata
 if typing.TYPE_CHECKING:  # pragma: nocover
     from charms.openfga_k8s.v1.openfga import OpenfgaProviderAppData, OpenFGARequires
     from charms.smtp_integrator.v0.smtp import SmtpRelationData, SmtpRequires
+    from charms.squid_forward_proxy.v0.http_proxy import ProxyConfig
 
     from paas_charm.databases import PaaSDatabaseRelationData, PaaSDatabaseRequires
+    from paas_charm.http_proxy import PaaSHttpProxyRequirer
     from paas_charm.oauth import PaaSOAuthRelationData, PaaSOAuthRequirer
     from paas_charm.rabbitmq import PaaSRabbitMQRelationData, RabbitMQRequires
     from paas_charm.redis import PaaSRedisRelationData, PaaSRedisRequires
@@ -101,6 +107,7 @@ class CharmState:  # pylint: disable=too-many-instance-attributes
 
         Raises:
             CharmConfigInvalidError: If some parameter in invalid.
+            RelationDataError: When relation data is either unavailable, invalid or not usable.
         """
         user_defined_config = {
             k.replace("-", "_"): v
@@ -178,9 +185,24 @@ class CharmState:  # pylint: disable=too-many-instance-attributes
                     if integration_requirers.oauth
                     else None
                 ),
+                http_proxy=(
+                    integration_requirers.http_proxy.fetch_proxies()
+                    if (
+                        integration_requirers.http_proxy
+                        and integration_requirers.http_proxy.model.get_relation("http-proxy")
+                    )
+                    else None
+                ),
             )
         except InvalidRelationDataError as exc:
-            raise CharmConfigInvalidError(f"Invalid {exc.relation} relation data.") from exc
+            raise RelationDataError(
+                f"Invalid {exc.relation} relation data.", relation=exc.relation
+            ) from exc
+        except RelationDataError as exc:
+            raise RelationDataError(
+                f"{exc.relation} relation data is either unavailable, invalid or not usable.",
+                relation=exc.relation,
+            ) from exc
         peer_fqdns = None
         if secret_storage.is_initialized and (
             peer_unit_fqdns := secret_storage.get_peer_unit_fdqns()
@@ -203,16 +225,20 @@ class CharmState:  # pylint: disable=too-many-instance-attributes
         )
 
     @property
-    def proxy(self) -> "ProxyConfig":
+    def proxy(self) -> "PaasProxyConfig":
         """Get charm proxy information from juju charm environment.
 
         Returns:
-            charm proxy information in the form of `ProxyConfig`.
+            charm proxy information in the form of `PaasProxyConfig`.
         """
-        http_proxy = os.environ.get("JUJU_CHARM_HTTP_PROXY")
-        https_proxy = os.environ.get("JUJU_CHARM_HTTPS_PROXY")
+        if self.integrations.http_proxy:
+            http_proxy = self.integrations.http_proxy.http_proxy
+            https_proxy = self.integrations.http_proxy.https_proxy
+        else:
+            http_proxy = os.environ.get("JUJU_CHARM_HTTP_PROXY")
+            https_proxy = os.environ.get("JUJU_CHARM_HTTPS_PROXY")
         no_proxy = os.environ.get("JUJU_CHARM_NO_PROXY")
-        return ProxyConfig(
+        return PaasProxyConfig(
             http_proxy=http_proxy if http_proxy else None,
             https_proxy=https_proxy if https_proxy else None,
             no_proxy=no_proxy,
@@ -277,6 +303,7 @@ class IntegrationRequirers:  # pylint: disable=too-many-instance-attributes
         smtp: Smtp requirer object.
         openfga: OpenFGA requirer object.
         oauth: PaaSOAuthRequirer object.
+        http_proxy: PaaSHttpProxyRequirer object.
     """
 
     databases: dict[str, "PaaSDatabaseRequires"]
@@ -288,6 +315,7 @@ class IntegrationRequirers:  # pylint: disable=too-many-instance-attributes
     tracing: "PaaSTracingEndpointRequirer | None" = None
     smtp: "SmtpRequires | None" = None
     oauth: "PaaSOAuthRequirer | None" = None
+    http_proxy: "PaaSHttpProxyRequirer | None" = None
 
 
 @dataclass
@@ -306,6 +334,7 @@ class IntegrationsState:  # pylint: disable=too-many-instance-attributes
         smtp: SMTP parameters.
         tracing: Tracing relation data.
         oauth: OAuth relation data.
+        http_proxy: HTTP proxy relation data.
     """
 
     databases_relation_data: dict[str, "PaaSDatabaseRelationData"] = field(default_factory=dict)
@@ -317,9 +346,10 @@ class IntegrationsState:  # pylint: disable=too-many-instance-attributes
     smtp: "SmtpRelationData | None" = None
     tracing: "PaaSTracingRelationData | None" = None
     oauth: "PaaSOAuthRelationData | None" = None
+    http_proxy: "ProxyConfig | None" = None
 
 
-class ProxyConfig(BaseModel):
+class PaasProxyConfig(BaseModel):
     """Configuration for network access through proxy.
 
     Attributes:
@@ -328,8 +358,8 @@ class ProxyConfig(BaseModel):
         no_proxy: Comma separated list of hostnames to bypass proxy.
     """
 
-    http_proxy: str | None = Field(default=None, pattern="https?://.+")
-    https_proxy: str | None = Field(default=None, pattern="https?://.+")
+    http_proxy: str | None = Field(default=None)
+    https_proxy: str | None = Field(default=None)
     no_proxy: typing.Optional[str] = None