+++ src/paas_charm/exceptions.py
@@ -28,11 +28,30 @@ class MissingCharmLibraryError(Exception):
     """Raised when a required charm library is missing."""
 
 
-class InvalidRelationDataError(Exception):
-    """Represents an invalid relation data.
+class RelationDataError(Exception):
+    """Raised when relation data is either unavailable, invalid or not usable.
 
-    Attributes:
-        relation: The name of the relation with invalid data.
+    Attrs:
+        relation: The name of the relation with error.
     """
 
-    relation: str
+    relation: str | None = None
+
+    def __init__(self, message: str, relation: str | None = None):
+        """Initialize a new instance of the RelationDataError exception.
+
+        Args:
+            message: Explanation of the error.
+            relation: The name of the relation with error.
+
+        Raises:
+            ValueError: If relation is not provided or defined on the class.
+        """
+        self.relation = relation or getattr(self, "relation", None)
+        if not self.relation:
+            raise ValueError(f"{self.__class__.__name__} requires a 'relation' to be set.")
+        super().__init__(message)
+
+
+class InvalidRelationDataError(RelationDataError):
+    """Raised when a relation data is invalid."""